completed

+++ MasaiSchoolCloneBackend/src/main/java/com/masaischoolclone/MasaiSchoolClone/controller/AssignmentController.java
@@ -1,7 +1,12 @@
 package com.masaischoolclone.MasaiSchoolClone.controller;
 
 import com.masaischoolclone.MasaiSchoolClone.entity.Assignment;
+import com.masaischoolclone.MasaiSchoolClone.entity.User;
+import com.masaischoolclone.MasaiSchoolClone.security.JwtTokenProvider;
 import com.masaischoolclone.MasaiSchoolClone.service.AssignmentService;
+import com.masaischoolclone.MasaiSchoolClone.service.UserService;
+import com.masaischoolclone.MasaiSchoolClone.utility.Common;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,12 +21,28 @@ public class AssignmentController {
     @Autowired
     private AssignmentService assignmentService;
 
-    @PostMapping("/create/{courseId}/{lectureId}")
-    public ResponseEntity<Assignment> assignmentCreate(@PathVariable Integer courseId, @PathVariable Integer lectureId, @RequestBody Assignment assignment){
+    @Autowired
+    private UserService userService;
+
+    @Autowired
+    private JwtTokenProvider jwtTokenProvider;
+
+    @PostMapping("/create/{userId}/{courseId}/{lectureId}")
+    public ResponseEntity<Assignment> assignmentCreate(@PathVariable Integer userId,@PathVariable Integer courseId,
+                                                       @PathVariable Integer lectureId,
+                                                       @RequestBody Assignment assignment,
+    HttpServletRequest request){
         try {
 
-            Assignment assignment1 = assignmentService.assignmentCreate(courseId,lectureId,assignment);
-            return new ResponseEntity<>(assignment1,HttpStatus.CREATED);
+            String userName = Common.getUserNameFromRequest(request,jwtTokenProvider);
+            User user = userService.getUser(userId);
+            if(user.getUsername().equals(userName) || user.getEmail().equals(userName)){
+                Assignment assignment1 = assignmentService.assignmentCreate(courseId,lectureId,assignment);
+                return new ResponseEntity<>(assignment1,HttpStatus.CREATED);
+            }else{
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
+
         } catch (Exception e) {
 
             e.printStackTrace();
@@ -30,11 +51,45 @@ public class AssignmentController {
         }
     }
 
-    @GetMapping("/fetch-all/{courseId}")
-    ResponseEntity<Set<Assignment>> getAssignmentList(@PathVariable Integer courseId){
+    @GetMapping("/fetch-all/{userId}/{courseId}")
+    ResponseEntity<Set<Assignment>> getAssignmentList(@PathVariable Integer userId,
+                                                      @PathVariable Integer courseId,
+                                                      HttpServletRequest request){
+//        try {
+
+            String userName = Common.getUserNameFromRequest(request,jwtTokenProvider);
+            User user = userService.getUser(userId);
+            if(user.getUsername().equals(userName) || user.getEmail().equals(userName)){
+                return ResponseEntity.ok(assignmentService.getAssignementList(courseId));
+            }else{
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
+
+
+//        } catch (Exception e) {
+//
+//            e.printStackTrace();
+//
+//            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+//
+//        }
+    }
+
+    @GetMapping("/fetch-all/{userId}/{courseId}/{lectureId}")
+    ResponseEntity<Set<Assignment>> getAssignmentListL(@PathVariable Integer userId,
+                                                       @PathVariable Integer courseId,
+                                                       @PathVariable Integer lectureId,
+                                                       HttpServletRequest request){
         try {
 
-            return ResponseEntity.ok(assignmentService.getAssignementList(courseId));
+            String userName = Common.getUserNameFromRequest(request,jwtTokenProvider);
+            User user = userService.getUser(userId);
+            if(user.getUsername().equals(userName) || user.getEmail().equals(userName)){
+                return ResponseEntity.ok(assignmentService.getAssignmentList(courseId,lectureId));
+            }else{
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
+
         } catch (Exception e) {
 
             e.printStackTrace();
@@ -44,25 +99,21 @@ public class AssignmentController {
         }
     }
 
-    @GetMapping("/fetch-all/{courseId}/{lectureId}")
-    ResponseEntity<Set<Assignment>> getAssignmentList(@PathVariable Integer courseId,@PathVariable Integer lectureId){
+    @PutMapping("/update/{userId}/{updateId}")
+    ResponseEntity<Assignment> updateAssignment(@PathVariable Integer userId,@PathVariable Integer updateId,
+                                                @RequestBody Assignment updatedAssignment,
+                                                HttpServletRequest request){
         try {
 
-            return ResponseEntity.ok(assignmentService.getAssignmentList(courseId,lectureId));
-        } catch (Exception e) {
+            String userName = Common.getUserNameFromRequest(request,jwtTokenProvider);
+            User user = userService.getUser(userId);
+            if(user.getUsername().equals(userName) || user.getEmail().equals(userName)){
+                return ResponseEntity.ok(assignmentService.updateAssignment(updateId,updatedAssignment));
+            }else{
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
 
-            e.printStackTrace();
 
-            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
-
-        }
-    }
-
-    @PutMapping("/update/{updateId}")
-    ResponseEntity<Assignment> updateAssignment(@PathVariable Integer updateId,@RequestBody Assignment updatedAssignment){
-        try {
-
-            return ResponseEntity.ok(assignmentService.updateAssignment(updateId,updatedAssignment));
         } catch (Exception e) {
 
             e.printStackTrace();
@@ -73,11 +124,18 @@ public class AssignmentController {
 
     }
 
-    @DeleteMapping("/delete/{assignmentId}")
-    ResponseEntity<Integer> deleteAssignment(@PathVariable Integer assignmentId){
+    @DeleteMapping("/delete/{userId}/{assignmentId}")
+    ResponseEntity<Integer> deleteAssignment(@PathVariable Integer userId,@PathVariable Integer assignmentId,HttpServletRequest request){
         try {
 
-            return ResponseEntity.ok(assignmentService.deleteAssignment(assignmentId));
+            String userName = Common.getUserNameFromRequest(request,jwtTokenProvider);
+            User user = userService.getUser(userId);
+            if(user.getUsername().equals(userName) || user.getEmail().equals(userName)){
+                return ResponseEntity.ok(assignmentService.deleteAssignment(assignmentId));
+            }else{
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
+
         } catch (Exception e) {
 
             e.printStackTrace();
@@ -87,11 +145,18 @@ public class AssignmentController {
         }
     }
 
-    @GetMapping("/fetch/{assignmentId}")
-    ResponseEntity<Assignment> getAssignment(@PathVariable Integer assignmentId){
+    @GetMapping("/fetch/{userId}/{assignmentId}")
+    ResponseEntity<Assignment> getAssignment(@PathVariable Integer userId,@PathVariable Integer assignmentId,HttpServletRequest request){
         try {
 
-            return ResponseEntity.ok(assignmentService.getAssignment(assignmentId));
+            String userName = Common.getUserNameFromRequest(request,jwtTokenProvider);
+            User user = userService.getUser(userId);
+            if(user.getUsername().equals(userName) || user.getEmail().equals(userName)){
+                return ResponseEntity.ok(assignmentService.getAssignment(assignmentId));
+            }else{
+                return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+            }
+
         } catch (Exception e) {
 
             e.printStackTrace();
